Break karma.js into multiple modules
karma.js had grown too large and unwieldy. Needed to be broken down.
