+++ lib/conversation.js
@@ -1,27 +1,22 @@
 const fs = require('fs');
-const winston = require('winston');
-const setup = require('./setup.js');
 const {
   facebookUtils,
-  facebook,
+
   services,
   config,
-  http,
+
   translate: t,
   localeUtils,
 } = require('borq');
+
 const {
   nextConversation,
   goto,
   generateButtonTemplate,
   generateQuickReply,
 } = facebookUtils;
-const idString = config.onaFormIds.default;
-const {controller} = facebook;
-const karma = facebook.controller.spawn({});
-let lang = config.defaultLanguage;
-setup(karma);
 
+const idString = config.onaFormIds.default;
 
 /**
 * Create an object for the withWhom section
@@ -344,161 +339,4 @@ function karmaConversation(err, convo, {uuid, name, language}) {
   });
 }
 
-
-/**
-* Fetch facebook user profile get the reponent's language and name
-* No return statement.
-* We use it to start the conversation and get the user profile from Facebook.
-* @param {object} bot A bot instance created by the controller
-* @param {object} message a message object also created by the controller
-* @param {string} newLanguage An ISO6392 language code string
-*/
-function prepareConversation(bot, message, newLanguage) {
-  const {user} = message;
-  if (newLanguage) {
-    services.updateUser({urn: `facebook:${user}`},
-                                   {language: newLanguage})
-      .then((rapidProContact) => {
-        bot.startConversation(message, (err, convo) => {
-          karmaConversation(err, convo, rapidProContact);
-        });
-      })
-      .catch((reason) =>
-             http.genericCatchRejectedPromise(
-               'Failed to updateRapidProContact in prepareConversation:' +
-                 ` ${reason}`));
-  } else {
-    services.getUser({urn: `facebook:${user}`})
-      .then(({results: [rapidProContact]}) =>
-            bot.startConversation(message, (err, convo) => {
-              karmaConversation(err, convo, rapidProContact);
-      }))
-      .catch((reason) => createUserAndStartConversation(message, bot));
-  }
-}
-
-/**
-* Get the region for a specific timezone
-* @param {Integer} timezone The timezone of the user runs from 0 to 12
-* @return {String} the region that the user is in
-*/
-function regionByTimeZone(timezone) {
-  if (timezone < 0) {
-    return 'por';
-  } else if (timezone > 3) {
-    return 'ind';
-  } else {
-    return 'default';
-  }
-}
-
-/**
-* First try locale, if it fails try timezone, if that fails use default
-* @param {String} locale of the user from FB
-* @param {String} timezone of the user from FB
-* @return {String} language as a ISO6392 string
-*/
-function pickLanguage({locale, timezone}) {
-  if (locale) {
-    const lang = localeUtils.extractLanguageFromLocale(locale) ||
-          config.defaultLanguage;
-    return localeUtils.lookupISO6392(lang);
-  } else {
-    const region = regionByTimeZone(timezone);
-    const lang = 'default' ? config.defaultLanguage : region;
-    return localeUtils.lookupISO6392(lang);
-  }
-}
-
-/**
-* Create a karma user and start a conversation with them
-* @param {object} message a message object also created by the controller
-* @param {object} bot a bot instance created by the controller
-*/
-function createUserAndStartConversation(message, bot) {
-  services.getFacebookProfile(message.user)
-    .then((profile) => {
-       const {first_name: firstName,
-              last_name: lastName,
-              profile_pic,
-              locale,
-              timezone,
-              gender,
-              is_payment_enabled} = profile;
-      const region = regionByTimeZone(timezone);
-      const referrer = message.referral ? message.referral.ref : 'none';
-      services.createUser(`${firstName} ${lastName}`,
-                          pickLanguage(profile),
-                          [`facebook:${message.user}`],
-                          [config.rapidproGroups[region]],
-                          {
-                            profile_pic,
-                            locale,
-                            timezone,
-                            gender,
-                            is_payment_enabled,
-                            referrer,
-                          })
-        .then((rapidProContact) => {
-          bot.startConversation(message, (err, convo) => {
-            karmaConversation(err, convo, rapidProContact);
-          });
-        })
-        .catch((reason) => http.genericCatchRejectedPromise(
-          `Failed createUser: ${reason}`));
-    })
-    .catch((reason) => http.genericCatchRejectedPromise(
-      'Failed to fetch Facebook profile' +
-        `in createUserAndStartConversation: ${reason}`));
-}
-
-/* Listeners */
-controller.on('facebook_postback', (bot, message) => {
-  const {payload} = message;
-  if (['restart', 'get_started'].includes(payload)) {
-    prepareConversation(bot, message);
-  } else if (['switch_pt', 'switch_en', 'switch_in'].includes(payload)) {
-    lang = payload.split('_')[1];
-    prepareConversation(bot, message, localeUtils.lookupISO6392(lang));
-  } else if (['quit'].includes(payload)) {
-    bot.reply(message, t(`${lang}:utils.quitMessage`));
-  } else if (['opt_out'].includes(payload)) {
-    services.deleteUser(message.user, Object.values(config.deletedUserGroups));
-    bot.reply(message, t(`${lang}:utils.quitMessage`));
-  }
-});
-
-controller.on('facebook_referral', (bot, message) => {
-  createUserAndStartConversation(message, bot);
-});
-
-controller.hears(['help',
-                   '👋',
-                   'hello',
-                   'halo',
-                   'hi',
-                   'oi',
-                   'hai',
-                   'membantu',
-                   'socorro',
-                   'olá'], 'message_received', (bot, message) => {
-                    bot.reply(message,
-                              generateButtonTemplate(
-                                t(`${lang}:utils.helpMessage`),
-                                null,
-                                [{
-                                  title: t(`${lang}:yes`),
-                                  payload: 'restart',
-                                }, {
-                                  title: t(`${lang}:no`),
-                                  payload: 'quit',
-                                }]));
-                   });
-
-controller.hears([''], 'message_received', (bot, message) => {});
-
-controller.hears([/([a-z])\w+/gi],
-               'message_received',
-               function(bot, message) {
-                 bot.reply(message, t(`${lang}:utils.idkw`));
-});
+module.exports = karmaConversation;
