+++ lib/setup.js
@@ -9,15 +9,23 @@ const {
 } = require('borq');
 const {
   generateButtonTemplate,
-  sendMessage
+  sendMessage,
 } = facebookUtils;
 
-function setup (bot) {
-
+/**
+* Holder doc
+* @param {object} bot -
+*/
+function setup(bot) {
+  /**
+  * Holder doc
+  * @param {object} contact -
+  */
   function sendGreeting({urn, contact_name, contact}) {
     services.getUser({urn: urn})
       .then(({results: [{language}]}) => {
-        let lang = localeUtils.lookupISO6391(language) || config.defaultLanguage;
+        let lang =
+        localeUtils.lookupISO6391(language) || config.defaultLanguage;
         let facebookId = urn.split(':')[1];
         sendMessage(bot, facebookId, (err, convo) => {
           convo.say(generateButtonTemplate(
@@ -99,7 +107,7 @@ function setup (bot) {
       webview_height_ratio: 'full',
     },
   ],
-},{
+}, {
   locale: 'pt_BR',
   composer_input_disabled: false,
   call_to_actions: [
@@ -146,7 +154,7 @@ function setup (bot) {
       webview_height_ratio: 'full',
     },
   ],
-},{
+}, {
   locale: 'id_ID',
   composer_input_disabled: false,
   call_to_actions: [
